requirements.txt
openai==1.5.0
python-dotenv~=1.0.0
keyboard~=0.13.5
psutil~=5.9.4
colorama~=0.4.6